Refactor wallet connection to use a modal component
