+++ src/app/app.html
@@ -104,3 +104,4 @@
 </div>
 
 <app-transaction-history />
+<app-wallet-modal />
